Group PI05, GR00T, and Nemotron under exporters/models.
Each family now has spec, helpers, and patches in one folder, with shared vision/language/action-step code in models/common.
