+++ py/torch_tensorrt/hf/exporters/__init__.py
@@ -1,14 +1,20 @@
 from torch_tensorrt.hf.exporters.config import EdgeConfig
 from torch_tensorrt.hf.exporters.exporter import EdgeExporter
+from torch_tensorrt.hf.exporters.models.groot.spec import (  # noqa: F401
+    GrootSpec as _GrootSpec,
+)
+from torch_tensorrt.hf.exporters.models.nemotron.spec import (  # noqa: F401
+    NemotronSpec as _NemotronSpec,
+)
+from torch_tensorrt.hf.exporters.models.pi05.spec import (  # noqa: F401
+    Pi05Spec as _Pi05Spec,
+)
 from torch_tensorrt.hf.exporters.spec import (
     ComponentBundle,
     EdgeSpec,
     get_edge_spec,
     register_edge_spec,
 )
-from torch_tensorrt.hf.exporters.specs import groot as _groot  # noqa: F401
-from torch_tensorrt.hf.exporters.specs import nemotron as _nemotron  # noqa: F401
-from torch_tensorrt.hf.exporters.specs import pi05 as _pi05  # noqa: F401
 
 __all__ = [
     "ComponentBundle",
