+++ py/torch_tensorrt/hf/exporters/models/__init__.py
@@ -1,0 +1,5 @@
+"""Model families. Importing this package registers EdgeSpecs."""
+
+from torch_tensorrt.hf.exporters.models.groot import spec as _groot  # noqa: F401
+from torch_tensorrt.hf.exporters.models.nemotron import spec as _nemotron  # noqa: F401
+from torch_tensorrt.hf.exporters.models.pi05 import spec as _pi05  # noqa: F401
