+++ py/torch_tensorrt/hf/exporters/models/groot/helpers.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from typing import Any
 
 import torch
+import torch.nn as nn
 
 GROOT_EMBODIMENT_MAPPING = {
     "new_embodiment": 31,
@@ -24,4 +25,15 @@ def make_embodiment_id(
         GROOT_EMBODIMENT_MAPPING.get(embodiment_tag, 0),
         dtype=dtype,
         device=device,
+    )
+
+
+def _groot(model: nn.Module) -> nn.Module:
+    if hasattr(model, "_groot_model"):
+        return model._groot_model
+    backbone = getattr(model, "backbone", None)
+    if backbone is not None and hasattr(backbone, "eagle_model"):
+        return model
+    raise RuntimeError(
+        "GR00T spec expected GrootPolicy or a module with backbone.eagle_model"
     )
