+++ py/torch_tensorrt/hf/exporters/models/groot/patches.py
@@ -1,8 +1,9 @@
+from __future__ import annotations
+
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-from lerobot.policies.pi05.modeling_pi05 import create_sinusoidal_pos_embedding
-from torch_tensorrt.hf.exporters.prefix_cache import PrefixKVCache
+from torch_tensorrt.hf.exporters.models.common.patches import ActionStepEncoderPatch
 
 
 class TRTFixedCategorySpecificLinearPatch(nn.Module):
@@ -149,92 +150,6 @@ class TRTDynamicGrootActionEncoderPatch(nn.Module):
         return self.W3(hidden, embodiment_id)
 
 
-class ActionStepEncoderPatch(nn.Module):
-    """Base contract for model-specific action-step encoding.
-
-    Subclasses implement forward() to turn noisy actions, timestep, and
-    model-specific context tensors into the args/kwargs consumed by the action
-    expert and velocity decoder. The default helpers cover common expert output
-    and velocity shapes, while model-specific encoders can override them.
-    """
-
-    def get_action_hidden(self, expert_out, output_tokens: int):
-        # Default path for experts that return either a standard model output
-        # with last_hidden_state, a raw hidden-state tensor, or a tuple/list whose
-        # first item is the hidden-state tensor.
-        hidden = (
-            expert_out.last_hidden_state
-            if hasattr(expert_out, "last_hidden_state")
-            else expert_out
-        )
-
-        if isinstance(hidden, (tuple, list)):
-            hidden = hidden[0]
-
-        return hidden[:, -output_tokens:]
-
-    def process_velocity(self, velocity):
-        # Default path for models whose decoder already returns the final action
-        # velocity shape. Override for models that need reshaping or cropping.
-        return velocity
-
-
-class StaticActionVelocityStepPatch(nn.Module):
-    """One static denoising step shared by VLA action diffusion modules.
-
-    The model-specific step_encoder owns the messy part: converting noisy
-    actions, timestep, and context tensors into the exact action_expert call.
-    This wrapper only runs the expert, selects action-token hidden states, and
-    decodes those hidden states into a velocity update.
-    """
-
-    def __init__(
-        self,
-        *,
-        step_encoder: ActionStepEncoderPatch,
-        action_expert: nn.Module,
-        velocity_decoder: nn.Module,
-        output_tokens: int,
-        cast_hidden_fp32: bool = True,
-    ):
-        super().__init__()
-        self.step_encoder = step_encoder
-        self.action_expert = action_expert
-        self.velocity_decoder = velocity_decoder
-        self.output_tokens = int(output_tokens)
-        self.cast_hidden_fp32 = cast_hidden_fp32
-
-    def forward(self, x_t, timestep, *inputs):
-        # Build the action expert inputs and any decoder-specific side inputs.
-        expert_args, expert_kwargs, decoder_args, decoder_kwargs = self.step_encoder(
-            x_t,
-            timestep,
-            *inputs,
-        )
-
-        # Run the model-specific action expert: Gemma expert, DiT, etc.
-        expert_out = self.action_expert(*expert_args, **expert_kwargs)
-
-        # Most experts return last_hidden_state, but some wrappers return tuples
-        # or need custom suffix-token selection.
-        action_hidden = self.step_encoder.get_action_hidden(
-            expert_out,
-            self.output_tokens,
-        )
-
-        if self.cast_hidden_fp32:
-            action_hidden = action_hidden.to(dtype=torch.float32)
-
-        # Project action-token hidden states back to action-space velocity.
-        velocity = self.velocity_decoder(
-            action_hidden,
-            *decoder_args,
-            **decoder_kwargs,
-        )
-
-        return self.step_encoder.process_velocity(velocity)
-
-
 class GrootDiTStepEncoderPatch(ActionStepEncoderPatch):
     def __init__(self, action_head, embodiment_id: torch.Tensor | None = None):
         super().__init__()
@@ -292,47 +207,22 @@ class GrootDiTStepEncoderPatch(ActionStepEncoderPatch):
         return expert_args, expert_kwargs, decoder_args, decoder_kwargs
 
 
-class PI05PrefixKVStepEncoderPatch(ActionStepEncoderPatch):
-    """PI05 suffix embed + AdaRMS cond, consumed by Gemma action expert."""
+class ContextProjectionPatch(nn.Module):
+    """eagle_linear -> vlln -> vl_self_attention (matches eager context path)."""
 
-    def __init__(self, core):
+    def __init__(self, eagle_linear, vlln, vl_self_attention):
         super().__init__()
-        self.action_in_proj = core.action_in_proj
-        self.time_mlp_in = core.time_mlp_in
-        self.time_mlp_out = core.time_mlp_out
-        self.config = core.config
-        self.hidden_size = core.action_in_proj.out_features
+        self.eagle_linear = eagle_linear
+        self.vlln = vlln
+        self.vl_self_attention = vl_self_attention
 
-    def forward(
-        self,
-        x_t,
-        timestep,
-        prefix_k,
-        prefix_v,
-        position_ids,
-        attention_mask,
-    ):
-        suffix_embs = self.action_in_proj(x_t)
+    def forward(self, hidden_states: torch.Tensor):
+        context_embs = self.eagle_linear(hidden_states)
 
-        time_emb = create_sinusoidal_pos_embedding(
-            timestep,
-            self.hidden_size,
-            min_period=self.config.min_period,
-            max_period=self.config.max_period,
-            device=timestep.device,
-        ).to(dtype=suffix_embs.dtype)
+        vlln_weight = getattr(self.vlln, "weight", None)
+        if vlln_weight is not None:
+            context_embs = context_embs.to(dtype=vlln_weight.dtype)
 
-        adarms_cond = self.time_mlp_in(time_emb)
-        adarms_cond = F.silu(adarms_cond)
-        adarms_cond = self.time_mlp_out(adarms_cond)
-        adarms_cond = F.silu(adarms_cond)
-
-        expert_kwargs = {
-            "inputs_embeds": suffix_embs,
-            "attention_mask": attention_mask,
-            "position_ids": position_ids,
-            "past_key_values": PrefixKVCache(prefix_k, prefix_v),
-            "use_cache": False,
-            "adarms_cond": adarms_cond,
-        }
-        return (), expert_kwargs, (), {}
+        context_embs = self.vlln(context_embs)
+        context_embs = self.vl_self_attention(context_embs)
+        return context_embs
