--- py/torch_tensorrt/hf/exporters/specs/__init__.py
@@ -1,5 +1,0 @@
-"""Family specs. Import is a side effect: ``@register_edge_spec`` attaches."""
-
-from torch_tensorrt.hf.exporters.specs import groot, nemotron, pi05
-
-__all__ = ["groot", "nemotron", "pi05"]
