+++ setup.py
@@ -618,10 +618,12 @@ dynamo_packages = [
     "torch_tensorrt.executorch",
     "torch_tensorrt.hf",
     "torch_tensorrt.hf.exporters",
-    "torch_tensorrt.hf.exporters.helpers",
-    "torch_tensorrt.hf.exporters.patches",
+    "torch_tensorrt.hf.exporters.models",
+    "torch_tensorrt.hf.exporters.models.common",
+    "torch_tensorrt.hf.exporters.models.groot",
+    "torch_tensorrt.hf.exporters.models.nemotron",
+    "torch_tensorrt.hf.exporters.models.pi05",
     "torch_tensorrt.hf.exporters.plugin",
-    "torch_tensorrt.hf.exporters.specs",
     "torch_tensorrt.runtime",
 ]
 
@@ -663,10 +665,12 @@ dynamo_package_dir = {
     "torch_tensorrt.executorch": "py/torch_tensorrt/executorch",
     "torch_tensorrt.hf": "py/torch_tensorrt/hf",
     "torch_tensorrt.hf.exporters": "py/torch_tensorrt/hf/exporters",
-    "torch_tensorrt.hf.exporters.helpers": "py/torch_tensorrt/hf/exporters/helpers",
-    "torch_tensorrt.hf.exporters.patches": "py/torch_tensorrt/hf/exporters/patches",
+    "torch_tensorrt.hf.exporters.models": "py/torch_tensorrt/hf/exporters/models",
+    "torch_tensorrt.hf.exporters.models.common": "py/torch_tensorrt/hf/exporters/models/common",
+    "torch_tensorrt.hf.exporters.models.groot": "py/torch_tensorrt/hf/exporters/models/groot",
+    "torch_tensorrt.hf.exporters.models.nemotron": "py/torch_tensorrt/hf/exporters/models/nemotron",
+    "torch_tensorrt.hf.exporters.models.pi05": "py/torch_tensorrt/hf/exporters/models/pi05",
     "torch_tensorrt.hf.exporters.plugin": "py/torch_tensorrt/hf/exporters/plugin",
-    "torch_tensorrt.hf.exporters.specs": "py/torch_tensorrt/hf/exporters/specs",
     "torch_tensorrt.runtime": "py/torch_tensorrt/runtime",
 }
